Delete the pager: mz_ore::pager, ColumnPager, ColumnMergeBatcher
