+++ misc/python/materialize/mzcompose/__init__.py
@@ -506,10 +506,7 @@ UNINTERESTING_SYSTEM_PARAMETERS = [
     "enable_column_paged_batcher",
     "enable_column_paged_batcher_spill",
     "column_paged_batcher_budget_fraction",
-    "column_paged_batcher_lz4",
-    "column_paged_batcher_swap_pageout",
-    "column_paged_batcher_use_pool",
-    "column_paged_batcher_pool_spill_threads",
+    "column_paged_batcher_spill_worker_count",
     "enable_upsert_paged_spill",
     "enable_lgalloc_eager_reclamation",
     "lgalloc_background_interval",
