+++ src/ore/src/lib.rs
@@ -63,16 +63,13 @@ pub mod option;
 #[cfg_attr(nightly_doc_features, doc(cfg(feature = "overflowing")))]
 #[cfg(feature = "overflowing")]
 pub mod overflowing;
-#[cfg_attr(nightly_doc_features, doc(cfg(all(feature = "pager", unix))))]
-#[cfg(all(feature = "pager", unix))]
-pub mod pager;
 #[cfg(not(target_family = "wasm"))]
 #[cfg(feature = "panic")]
 pub mod panic;
 pub mod path;
 pub mod permutations;
-#[cfg_attr(nightly_doc_features, doc(cfg(all(feature = "pager", unix))))]
-#[cfg(all(feature = "pager", unix))]
+#[cfg_attr(nightly_doc_features, doc(cfg(all(feature = "pool", unix))))]
+#[cfg(all(feature = "pool", unix))]
 pub mod pool;
 #[cfg(feature = "process")]
 pub mod process;
