+++ src/storage/src/storage_state.rs
@@ -844,11 +844,11 @@ impl<'w> Worker<'w> {
                     STORAGE_SERVER_MAINTENANCE_INTERVAL
                         .get(self.storage_state.storage_configuration.config_set());
 
-                // Gate the upsert source-stash's use of the column pager. The
-                // pager's budget pool, backend, and codec are the shared ones
-                // configured by compute's `apply_worker_config` (compute and
-                // storage run in the same process); storage only decides whether
-                // its stash participates, via its own dyncfg.
+                // Gate the upsert source-stash's chunk spilling. The buffer
+                // pool and its budget are the shared ones configured by
+                // compute's `apply_worker_config` (compute and storage run in
+                // the same process); storage only decides whether its stash
+                // participates, via its own dyncfg.
                 {
                     use mz_storage_types::dyncfgs::ENABLE_UPSERT_PAGED_SPILL;
 
@@ -856,9 +856,9 @@ impl<'w> Worker<'w> {
                         .get(self.storage_state.storage_configuration.config_set());
                     info!(
                         worker = self.timely_worker.index(),
-                        enabled, "upsert stash pager: applying gate",
+                        enabled, "upsert stash spill: applying gate",
                     );
-                    crate::upsert::upsert_stash_pager::set_enabled(enabled);
+                    crate::upsert::upsert_stash_spill::set_enabled(enabled);
                 }
             }
             InternalStorageCommand::StatisticsUpdate { sources, sinks } => self
