+++ src/timely-util/Cargo.toml
@@ -21,14 +21,6 @@ harness = false
 name = "columnar_merger"
 harness = false
 
-[[bench]]
-name = "columnar_merge_batcher"
-harness = false
-
-[[bench]]
-name = "column_pager"
-harness = false
-
 [dependencies]
 ahash.workspace = true
 bincode.workspace = true
@@ -39,8 +31,7 @@ differential-dataflow.workspace = true
 either.workspace = true
 futures-util.workspace = true
 lgalloc.workspace = true
-lz4_flex.workspace = true
-mz-ore = { path = "../ore", default-features = false, features = ["async", "process", "tracing", "test", "num-traits", "region", "differential-dataflow", "overflowing", "pager"] }
+mz-ore = { path = "../ore", default-features = false, features = ["async", "process", "tracing", "test", "num-traits", "region", "differential-dataflow", "overflowing", "pool"] }
 num-traits.workspace = true
 serde.workspace = true
 timely.workspace = true
@@ -58,7 +49,6 @@ criterion.workspace = true
 itertools.workspace = true
 proptest.workspace = true
 rand.workspace = true
-tempfile.workspace = true
 
 [features]
 default = ["mz-ore/default"]
