+++ src/timely-util/src/lib.rs
@@ -19,7 +19,6 @@ pub mod activator;
 pub mod antichain;
 pub mod builder_async;
 pub mod capture;
-pub mod column_pager;
 pub mod columnar;
 pub mod columnation;
 pub mod containers;
@@ -28,6 +27,7 @@ pub mod operator;
 pub mod order;
 pub mod pact;
 pub mod panic;
+pub mod pool_config;
 pub mod probe;
 pub mod reclock;
 pub mod replay;
